cleanup BackendCrumb mixin
And move the logic into the controller
