+++ ui/app/controllers/vault/cluster/secrets/backend/metadata.js
@@ -4,12 +4,24 @@
  */
 
 import Controller from '@ember/controller';
-import BackendCrumbMixin from 'vault/mixins/backend-crumb';
 import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
 
-export default class MetadataController extends Controller.extend(BackendCrumbMixin) {
+export default class MetadataController extends Controller {
   @action
   refreshModel() {
     this.send('refreshModel');
+  }
+
+  @tracked backend;
+  get backendCrumb() {
+    const backend = this.backend;
+
+    return {
+      label: backend,
+      text: backend,
+      path: 'vault.cluster.secrets.backend.list-root',
+      model: backend,
+    };
   }
 }
