+++ ui/app/controllers/vault/cluster/secrets/backend/show.js
@@ -4,9 +4,9 @@
  */
 
 import Controller, { inject as controller } from '@ember/controller';
-import BackendCrumbMixin from 'vault/mixins/backend-crumb';
+import { computed } from '@ember/object';
 
-export default Controller.extend(BackendCrumbMixin, {
+export default Controller.extend({
   backendController: controller('vault.cluster.secrets.backend'),
   queryParams: ['tab', 'version', 'type', 'itemType', 'page'],
   version: '',
@@ -19,6 +19,18 @@ export default Controller.extend(BackendCrumbMixin, {
     this.set('type', '');
     this.set('itemType', '');
   },
+
+  backendCrumb: computed('backend', function () {
+    const backend = this.backend;
+
+    return {
+      label: backend,
+      text: backend,
+      path: 'vault.cluster.secrets.backend.list-root',
+      model: backend,
+    };
+  }),
+
   actions: {
     refresh: function () {
       // closure actions don't bubble to routes,
